% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateEnsemblBiomart.R
\name{annotateEnsemblBiomart}
\alias{annotateEnsemblBiomart}
\title{Annotate Ensembl Gene IDs with Entrez GeneIDs, gene symbols, and names using Biomart}
\usage{
annotateEnsemblBiomart(ids, mart = ensemblMart("ensembl", species =
  "hsapiens"), multi = FALSE, orthologue = FALSE, multiOrth = FALSE)
}
\arguments{
\item{ids}{Ensembl Gene IDs}

\item{mart}{A Mart or LocalMart object}

\item{multi}{Logical, in case one ensembl Gene ID is mapped to multiple Entrez GeneIDs, should multiple IDs returned. Default: \code{FALSE}}

\item{orthologue}{Logical, currently ignored}

\item{multiOrth}{Logical, currently ignored}
}
\description{
Annotate Ensembl Gene IDs with Entrez GeneIDs, gene symbols, and names using Biomart
}
\examples{
\dontrun{
  myMart <- ensemblMart("ensembl", species="hsapiens")
  annotateEnsemblBiomart(c("ENSG00000142208"), mart=myMart)
  annotateEnsemblBiomart(c("ENSG00000142208",
    "ENSG00000105221",
    "ENSG00000117020"), mart=myMart)
}

}
