% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\docType{methods}
\name{createDBIConnection,EnsemblDBCredentials-method}
\alias{createDBIConnection,EnsemblDBCredentials-method}
\title{Creates a new DBI connection based on the given credential object}
\usage{
\S4method{createDBIConnection}{EnsemblDBCredentials}(object)
}
\arguments{
\item{object}{A EnsemblDBCrednetials object}
}
\description{
Creates a new DBI connection based on the given credential object
}
