% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\docType{class}
\name{EnsemblDBCredentials-class}
\alias{EnsemblDBCredentials-class}
\alias{EnsemblDBCredentials}
\title{This class is used to store credentials used to connect a MySQL / MariaDB fork of the Ensembl database}
\arguments{
\item{host}{The host name of the SQL database containing a copy of the Ensembl BioMart data}

\item{port}{The port to the SQL database}

\item{user}{The user name used to authenticate to the SQL database}

\item{passwd}{The database password used to connect to the SQL database}

\item{ensembl_version}{The version of the Ensembl data inside the target SQL database. Will be used as table suffix (e.g. for "ensembl_mart_92")}
}
\description{
This class is used to store credentials used to connect a MySQL / MariaDB fork of the Ensembl database
}
