% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marts.R
\name{ensemblMart}
\alias{ensemblMart}
\title{Construct a Mart object with the dataset of the given species}
\usage{
ensemblMart(mart = "ensembl", species = "hsapiens", ...)
}
\arguments{
\item{mart}{Character string, mart to be used, "ensembl" by default.}

\item{species}{Character string, species name to be used. Common names are accepted for commonly used species.}

\item{...}{Other parameters passed to \code{\link[biomaRt]{useMart}}}
}
\value{
A \code{Mart} object of the selected dataset
}
\description{
Construct a Mart object with the dataset of the given species
}
\examples{
\dontrun{
  humanMart <- ensemblMart(species="human")
  ratMart <- ensemblMart(species="rnorvegicus")
}
}
\seealso{
\code{\link{ensemblSpeciesName}}, \code{\link{ensemblMartDataset}}
}
