% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marts.R
\name{ensemblMartDataset}
\alias{ensemblMartDataset}
\title{Construct the Ensembl mart dataset name using species name}
\usage{
ensemblMartDataset(species = "hsapiens")
}
\arguments{
\item{species}{Character string, the species name (e.g. \code{H. sapiens}). Common names are accepted for commonly used species.}
}
\value{
The dataset name of the Ensembl mart of that species
}
\description{
Construct the Ensembl mart dataset name using species name
}
\examples{
ensemblMartDataset("hsapiens")
ensemblMartDataset("mmusculus")

}
\seealso{
\code{\link{ensemblSpeciesName}}
}
