% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marts.R
\name{ensemblSpeciesName}
\alias{ensemblSpeciesName}
\title{Get Ensembl species name}
\usage{
ensemblSpeciesName(species = "hsapiens")
}
\arguments{
\item{species}{Character string, either the common name of commonly used species (see below), or the species name in the form of \code{Homo sapiens} or \code{H. sapiens}}
}
\value{
The Ensembl species name in the form of \code{hsapiens}
}
\description{
Get Ensembl species name
}
\note{
Commonly used species include: human, mouse, rat, rabbit, chimpanzee, cyno (crab-eating monkey), 
dog, pig, chicken, and zebrafish
}
\examples{
## species names
ensemblSpeciesName("Homo sapiens")
ensemblSpeciesName("H.sapiens")
ensemblSpeciesName("H. sapiens")
ensemblSpeciesName("Gallus gallus")
## common names
ensemblSpeciesName("human")
ensemblSpeciesName("pig")
## in case the input name is already valid Ensembl name
ensemblSpeciesName("hsapiens")
}
