% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{extractInfoFromAttributeDescriptions}
\alias{extractInfoFromAttributeDescriptions}
\title{Used to extract XML attribute values from an list of AttributeDescription nodes}
\usage{
extractInfoFromAttributeDescriptions(nodes, name)
}
\arguments{
\item{nodes}{list of AttributeDescription XML nodes}

\item{name}{the name of the attribute we want to extract}
}
\value{
vector with XML attribute values or \code{NA} if the corresponding node doesn't have the requested XML attribute
}
\description{
Used to extract XML attribute values from an list of AttributeDescription nodes
}
