% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{extractListOfBioMartAttributes}
\alias{extractListOfBioMartAttributes}
\title{Used to extract the meta information about BioMart attributes from XML content}
\usage{
extractListOfBioMartAttributes(docRoot)
}
\arguments{
\item{docRoot}{the root XML document containg dataset meta information}
}
\value{
BioMart attribute meta information as \code{\link[data.table]{data.table}}
}
\description{
Used to extract the meta information about BioMart attributes from XML content
}
