% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{listLocalAttributes}
\alias{listLocalAttributes}
\title{Lists all available attributes for a selected dataset (see \code{\link[biomaRt]{listAttributes}}.}
\usage{
listLocalAttributes(mart)
}
\arguments{
\item{mart}{the BioMart meta object needed to execute the BioMart queries on a local database server}
}
\value{
a list of all available attributes as \code{\link[data.table]{data.table}}
}
\description{
Lists all available attributes for a selected dataset (see \code{\link[biomaRt]{listAttributes}}.
}
