% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{listLocalDatasets}
\alias{listLocalDatasets}
\title{Returns table of all available BioMart datasets (see \code{\link[biomaRt]{listDatasets}}}
\usage{
listLocalDatasets(conn)
}
\arguments{
\item{conn}{The \code{\linkS4class{DBIConnection}} connection object to the local Ensembl database or a instance of \code{\linkS4class{EnsemblDBCredentials}}}
}
\value{
a \code{\link[data.table]{data.table}}
}
\description{
Returns table of all available BioMart datasets (see \code{\link[biomaRt]{listDatasets}}
}
