% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{listLocalFilters}
\alias{listLocalFilters}
\title{Lists all available filters for a selected dataset (see \code{\link[biomaRt]{listFilters}}.}
\usage{
listLocalFilters(mart)
}
\arguments{
\item{mart}{the BioMart meta object needed to execute the BioMart queries on a local database server}
}
\value{
a list of all available filters as \code{\link[data.table]{data.table}}
}
\description{
Lists all available filters for a selected dataset (see \code{\link[biomaRt]{listFilters}}.
}
