% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{localFilterOptions}
\alias{localFilterOptions}
\title{Returns options of the given filters as list}
\usage{
localFilterOptions(filter, mart)
}
\arguments{
\item{filter}{the name of the filters as character vector}

\item{mart}{the BioMart meta object needed to execute the BioMart queries on a local database server}
}
\value{
a list of filter options available for each givn filter
}
\description{
Returns options of the given filters as list
}
