% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{localFilterType}
\alias{localFilterType}
\title{Returns the types of the given filters as vector}
\usage{
localFilterType(filter, mart)
}
\arguments{
\item{filter}{the name of the filters as character vector}

\item{mart}{the BioMart meta object needed to execute the BioMart queries on a local database server}
}
\value{
a vector of filter types
}
\description{
Returns the types of the given filters as vector
}
