% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marts.R
\name{rocheEnsemblMart}
\alias{rocheEnsemblMart}
\title{Construct a LocalMart object to access Roche internal Ensembl Mart}
\usage{
rocheEnsemblMart(species = "hsapiens", ensembl_version = 95,
  host = NULL, port = NULL, user = NULL, passwd = NULL)
}
\arguments{
\item{species}{Species to be used}

\item{ensembl_version}{Ensembl version to be used}

\item{host}{Character string, host server address}

\item{port}{Integer, port number}

\item{user}{Character string, user name}

\item{passwd}{Character string, password}
}
\value{
A \code{LocalMart} object which can be used to access Roche internal Ensembl Mart
}
\description{
Construct a LocalMart object to access Roche internal Ensembl Mart
}
\note{
If you are within Roche, the credentials can be obtained at [https://rochewiki.roche.com/confluence/display/BEDA/BIOMART+@+Roche](https://rochewiki.roche.com/confluence/display/BEDA/BIOMART+@+Roche).
}
\examples{
\dontrun{
  rocheEnsembleMart(species="hsapiens")
}
}
