% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\docType{methods}
\name{show,EnsemblDBCredentials-method}
\alias{show,EnsemblDBCredentials-method}
\title{Shows the objects states (without the password)}
\usage{
\S4method{show}{EnsemblDBCredentials}(object)
}
\arguments{
\item{object}{A EnsemblDBCrednetials object}
}
\description{
Shows the objects states (without the password)
}
