% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{useLocalMart}
\alias{useLocalMart}
\title{Creates a new mart object for the given database connection (see \code{\link[biomaRt]{useMart}}}
\usage{
useLocalMart(conn, dataset)
}
\arguments{
\item{conn}{The \code{\linkS4class{DBIConnection}} connection object to the local Ensembl database or an instance of \code{\linkS4class{EnsemblDBCredentials}}}

\item{dataset}{The name of the Ensembl dataset to query (e.g. "rnorvegicus_gene_ensembl")}
}
\value{
the mart object (regular list) with the connection object and internal attributes in it
}
\description{
Creates a new mart object for the given database connection (see \code{\link[biomaRt]{useMart}}
}
