% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{createDBIConnection}
\alias{createDBIConnection}
\title{Creates a new DBI connection based on the given credential object}
\usage{
createDBIConnection(object)
}
\arguments{
\item{object}{A DBIConnection object}
}
\description{
Creates a new DBI connection based on the given credential object
}
