% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_biomart.R
\name{getLocalBM}
\alias{getLocalBM}
\title{Executes a biomart query on the given mart object (see \code{\link[biomaRt]{getBM}}}
\usage{
getLocalBM(attributes, values, mart)
}
\arguments{
\item{attributes}{a vector of BioMart attributes}

\item{values}{the values used for the BioMart filters}

\item{mart}{the BioMart meta object needed to execute the BioMart queries on a local database server}
}
\value{
the mart object (regular list) with the connection object and internal attributes in it
}
\description{
Executes a biomart query on the given mart object (see \code{\link[biomaRt]{getBM}}
}
